backend/requirements.txt
flask
flask-cors
google-genai==1.10.0